I have started to turn the estry converter script into a class.
Also I have started to add it to component and model converter files so
that it can be used as a tool and be run from the terminal, but I have
an issue:

Issue in edit_fm_file for NetworkConverter1D in component_converter.py.
When I set self._dat equal to the _temp_dat the file wrapper doesn't
also change address pointer (in convert_model method in
TuflowModelConverter2D in model_converter.py) for the wrapper's fm_file
(which should be the same dat object but it isn't changing memory
address location - using id()). I also then tried only changing the
attributes (the ones that matter are sections, initial conditions, there
may be more...) but then the _write() method for the dat class was only
outputting the template dat class data (even if i do find out way this
method of fixing it is not as ideal as I dont know all the attributes
that have changed).